major overhaul, now only takes one argument

+++ xtest.sh
@@ -1,23 +1,31 @@
 #!/bin/dash
 
-#open like this
-#scriptname gallery_id number_of_pages nh_id
-#SAMPLE ./xtest.sh 530640 34 81391
+exit_here(){
+    echo "this doujin seems to not exist, if it exists leave an issue on github or message Zhāng Jūn 张军#8497"
+    exit 0
+}
 
-#create "sandbox" dir
-mkdir $3
-cd $3
+#gather doujin info
+doujin=$(curl -s "https://nhentai.net/api/gallery/$1")
+media_id=$(echo $doujin | jq -r '.media_id' || echo "please install jq")
+[ "${media_id}" = "null" ] && exit_here
+page_count=$(echo $doujin | jq '.images.pages | length ')
+id=$1
+name=$(echo $doujin | jq -r '.title.pretty')
 
-# local var $3
-name="$3"
+#create "doujin folder" dir
+mkdir $id
+cd $id
 
 #DL
-for I in `seq 1 $2`
+for I in `seq 1 $page_count`
 do
-    url="https://i.nhentai.net/galleries/$1/$I.jpg"
-    alturl="https://i.nhentai.net/galleries/$1/$I.png"
-    wget ${url} || wget ${alturl}
+    url="https://i.nhentai.net/galleries/$media_id/$I.jpg"
+    alturl="https://i.nhentai.net/galleries/$media_id/$I.png"
+    wget -q ${url} || wget -q ${alturl}
 done
+
+
 
 #oneliner lmao https://stackoverflow.com/a/3672383
 ls | awk '/^([0-9]+)\.jpg$/ { printf("%s %05d.jpg\n", $0, $1) }' | xargs -n2 mv
@@ -26,28 +34,28 @@ ls | awk '/^([0-9]+)\.png$/ { printf("%s %05d.png\n", $0, $1) }' | xargs -n2 mv
 #convert pictures
 convert(){
     mogrify -format jpg *.*
-    podofoimg2pdf "$name.pdf" -useimgsize *.jpg
-    rm *.png
+    podofoimg2pdf "$id.pdf" -useimgsize *.jpg
+    [ -f *.png ] && rm *.png
 }
 
 #PDFer
-podofoimg2pdf "$3.pdf" -useimgsize *.* || convert
+podofoimg2pdf "$id.pdf" -useimgsize *.* || convert
 
 #create file to remark nid
-echo "http://nhentai.net/g/$3/" > "$3.no"
-curl -s "https://nhentai.net/api/gallery/$3" | jq '.title.pretty' >> "$3.no" || echo "please install jq"
+echo "http://nhentai.net/g/$id/" > "$id.no"
+echo $name >> "$id.no"
 
 #zip it
-zip "$3.zip" *
+zip "$id.zip" *
 
 #get pwd
 _cwd="$PWD"
 cd ..
 
 #get top dir
-_top="$pwd"
+_top="$PWD"
 cd $_cwd
 
 #move zip top dir
-mv "$3.zip" ..
+mv "$id.zip" ..
